refactoring: define flask tests in single place and reuse them
