+++ tests/test_plugin_flask_blueprint.py
@@ -1,4 +1,3 @@
-import json
 from random import randint
 
 import pytest
@@ -18,6 +17,9 @@ from .common import (
     get_paths,
 )
 
+# import tests to execute
+from .flask_imports import *  # NOQA
+
 
 def before_handler(req, resp, err, _):
     if err:
@@ -32,7 +34,7 @@ def api_after_handler(req, resp, err, _):
     resp.headers["X-API"] = "OK"
 
 
-api = SpecTree("flask", before=before_handler, after=after_handler)
+api = SpecTree("flask", before=before_handler, after=after_handler, annotations=True)
 app = Blueprint("test_blueprint", __name__)
 
 
@@ -125,7 +127,7 @@ def user_address(name, address_id):
 
 @app.route("/api/no_response", methods=["GET", "POST"])
 @api.validate(
-    json=Query,
+    json=JSON,
 )
 def no_response():
     return {}
@@ -142,7 +144,8 @@ with flask_app.app_context():
 @pytest.fixture
 def client(request):
     parent_app = Flask(__name__)
-    parent_app.register_blueprint(app, url_prefix=request.param)
+    url_prefix = getattr(request, "param", None)
+    parent_app.register_blueprint(app, url_prefix=url_prefix)
     with parent_app.test_client() as client:
         yield client
 
@@ -150,74 +153,16 @@ def client(request):
 @pytest.mark.parametrize(
     ("client", "prefix"), [(None, ""), ("/prefix", "/prefix")], indirect=["client"]
 )
-def test_flask_validate(client, prefix):
+def test_blueprint_prefix(client, prefix):
     resp = client.get(prefix + "/ping")
     assert resp.status_code == 422
     assert resp.headers.get("X-Error") == "Validation Error"
 
     resp = client.get(prefix + "/ping", headers={"lang": "en-US"})
+    assert resp.status_code == 200
     assert resp.json == {"msg": "pong"}
     assert resp.headers.get("X-Error") is None
     assert resp.headers.get("X-Validation") == "Pass"
-
-    resp = client.post(prefix + "/api/user/flask")
-    assert resp.status_code == 422
-    assert resp.headers.get("X-Error") == "Validation Error"
-
-    client.set_cookie("flask", "pub", "abcdefg")
-    resp = client.post(
-        prefix + "/api/user/flask?order=1",
-        data=json.dumps(dict(name="flask", limit=10)),
-        content_type="application/json",
-    )
-    assert resp.status_code == 200, resp.json
-    assert resp.headers.get("X-Validation") is None
-    assert resp.headers.get("X-API") == "OK"
-    assert resp.json["name"] == "flask"
-    assert resp.json["score"] == sorted(resp.json["score"], reverse=True)
-
-    resp = client.post(
-        prefix + "/api/user/flask?order=0",
-        data=json.dumps(dict(name="flask", limit=10)),
-        content_type="application/json",
-    )
-    assert resp.json["score"] == sorted(resp.json["score"], reverse=False)
-
-
-@pytest.mark.parametrize(
-    ("client", "prefix"), [(None, ""), ("/prefix", "/prefix")], indirect=["client"]
-)
-def test_flask_skip_validation(client, prefix):
-    client.set_cookie("flask", "pub", "abcdefg")
-
-    resp = client.post(
-        f"{prefix}/api/user_skip/flask?order=1",
-        data=json.dumps(dict(name="flask", limit=10)),
-        content_type="application/json",
-    )
-    assert resp.status_code == 200, resp.json
-    assert resp.headers.get("X-Validation") is None
-    assert resp.headers.get("X-API") == "OK"
-    assert resp.json["name"] == "flask"
-    assert resp.json["x_score"] == sorted(resp.json["x_score"], reverse=True)
-
-
-@pytest.mark.parametrize(
-    ("client", "prefix"), [(None, ""), ("/prefix", "/prefix")], indirect=["client"]
-)
-def test_flask_return_model(client, prefix):
-    client.set_cookie("flask", "pub", "abcdefg")
-
-    resp = client.post(
-        f"{prefix}/api/user_model/flask?order=1",
-        data=json.dumps(dict(name="flask", limit=10)),
-        content_type="application/json",
-    )
-    assert resp.status_code == 200, resp.json
-    assert resp.headers.get("X-Validation") is None
-    assert resp.headers.get("X-API") == "OK"
-    assert resp.json["name"] == "flask"
-    assert resp.json["score"] == sorted(resp.json["score"], reverse=True)
 
 
 @pytest.fixture
@@ -260,71 +205,6 @@ def test_client_and_api(request):
 
     with flask_app.test_client() as test_client:
         yield test_client, api
-
-
-@pytest.mark.parametrize(
-    "test_client_and_api, expected_status_code",
-    [
-        pytest.param(
-            {"api_kwargs": {}, "endpoint_kwargs": {}},
-            422,
-            id="default-global-status-without-override",
-        ),
-        pytest.param(
-            {"api_kwargs": {}, "endpoint_kwargs": {"validation_error_status": 400}},
-            400,
-            id="default-global-status-with-override",
-        ),
-        pytest.param(
-            {"api_kwargs": {"validation_error_status": 418}, "endpoint_kwargs": {}},
-            418,
-            id="overridden-global-status-without-override",
-        ),
-        pytest.param(
-            {
-                "api_kwargs": {"validation_error_status": 400},
-                "endpoint_kwargs": {"validation_error_status": 418},
-            },
-            418,
-            id="overridden-global-status-with-override",
-        ),
-    ],
-    indirect=["test_client_and_api"],
-)
-def test_validation_error_response_status_code(
-    test_client_and_api, expected_status_code
-):
-    app_client, _ = test_client_and_api
-
-    resp = app_client.get("/ping")
-
-    assert resp.status_code == expected_status_code
-
-
-@pytest.mark.parametrize(
-    "test_client_and_api, expected_doc_pages",
-    [
-        pytest.param({}, ["redoc", "swagger"], id="default-page-templates"),
-        pytest.param(
-            {"api_kwargs": {"page_templates": {"custom_page": "{spec_url}"}}},
-            ["custom_page"],
-            id="custom-page-templates",
-        ),
-    ],
-    indirect=["test_client_and_api"],
-)
-def test_flask_doc(test_client_and_api, expected_doc_pages):
-    client, api = test_client_and_api
-
-    resp = client.get("/apidoc/openapi.json")
-    assert resp.json == api.spec
-
-    for doc_page in expected_doc_pages:
-        resp = client.get(f"/apidoc/{doc_page}/")
-        assert resp.status_code == 200
-
-        resp = client.get(f"/apidoc/{doc_page}")
-        assert resp.status_code == 308
 
 
 @pytest.mark.parametrize(
